KRW-maatlatten: Macrofauna categorieen

Aanduiding voor macrofauna taxa of een taxon voor een bepaald watertype behoort tot de positieve (P), kenmerkende (K) of negatieve (N) soorten. Deze tabel gaat over alle watertypen maar alleen over de maatlatten van 2018.

krw_mafa_categorie_2018

Format

Dataframe

biotaxon.naam Naam van het biotaxon, inclusief ondersoorten die niet expliciet in het maatlatdocument zijn opgenomen.

krwwatertype.code KRW-watertypecode

categorie Categorie van het taxon voor het betreffende watertype. Mogelijk categorieen zijn:

P - Positief taxon

K - Kenmerkend taxon

N - Negatief taxon

Source

Aquo-kit

Details

Macrofauna kan getoetst worden voor de Kaderrichtlijn Water d.m.v. zogenaamde maatlatten. De documentatie voor deze maatlatten kan gevonden worden op de website van STOWA. Het betreft:

Referenties en maatlatten voor natuurlijke watertypen voor de Kaderrichtlijn Water 2021-2027

Omschrijving MEP en maatlatten voor sloten en kanalen voor de Kaderrichtlijn Water 2021-2027

Note

In deze tabel zijn alleen de categorieën van de maatlat van 2018 opgenomen.

Metadata dataset

Naam: Macrofauna-categorieen

Oorspronkelijke Naam: Somparametersamenstelling, de macrofauna-categorieen zijn hier een subset van.

Bronhouder: IHW (voor Aquo-kit) en STOWA voor de onderliggende data

Herkomst: Aquo-kit https://www.aquo-kit.nl/

Auteurs: Diverse

Datum dataset: 2018

Datum opname dataset: 6-7-2019

Gebruiksrechten: Onbeperkt

Soortgroep: Macrofauna

Bewerkingen

De kolomnamen zijn omgezet naar lowercase.

See also

Examples

head(krw_mafa_categorie_2018)
#> # A tibble: 6 x 3
#>   biotaxon.naam          krwwatertype.code categorie
#>   <chr>                  <chr>             <chr>    
#> 1 Ablabesmyia longistyla M10               P        
#> 2 Ablabesmyia longistyla M1a               P        
#> 3 Ablabesmyia longistyla M1b               P        
#> 4 Ablabesmyia longistyla M2                P        
#> 5 Ablabesmyia longistyla M22               K        
#> 6 Ablabesmyia longistyla M3                P