KRW-toetsing macrofauna

Deze functie voert de KRW-toetsing voor macrofauna uit voor een of meerdere monsters. De toetsing is gebaseerd op de maatlatten uit het jaar 2018.

krw_mafa_ekr(df, krwwatertype.code, biotaxon.naam, waarde, verbose = FALSE)

Arguments

df
Een dataframe met macrofauna-data. Als het dataframe meerdere monsters bevat moet dit dataframe een grouped dataframe worden met de functie dplyr::group_by()

krwwatertype.code
Naam van de kolom met de KRW-watertype code

biotaxon.naam
Naam van de kolom met de taxonnamen.

waarde
Naam van de kolom met aantallen van het betreffende taxon.

verbose
Logical. Als TRUE dan worden behalve de EKR ook de onderliggende parameters en constanten in de uitvoer meegenomen. Deze zijn vooral relevant om e.e.a. te kunnen controleren.

Value

Een dataframe met groepkolommen (alleen indien van toepassing), het watertype en de macrofauna EKR.

Details

De kolomnamen van de kolommen met taxon, watertype en waarde kunnen zowel met als zonder quotes ("") worden ingevoerd door het gebruik van non-standard-evaluation (NSE).

Note

Het is van belang dat de biotaxonnamen in de TWN-lijst voorkomen. De KRW-watertypecode dient overeen te komen met de codes uit de codes uit de aquo-standaard.

De beoordeling van het watertype M30 is afhankelijk van de chlorideconcentratie. Deze functie geeft voor het watertype M30 3 EKR's met de daarbij behorende chloriderange. Hierdoor is het mogelijk om een toetsing uit te voeren zonder een chloride waarde. Ook is het mogelijk om naderhand de juist ekr te selecteren.

Deze functie is niet geschikt voor het watertypen R8.

Examples

if (FALSE) {

# Een enkel monster

krw_mafa_ekr(df_1_monster, watertype, taxon, aantallen)
krw_mafa_ekr(df_1_monster, "watertype", "taxon", "aantallen")

# Meerdere monsters in 1 dataframe

grouped_df <- group_by(df_meerdere_monsters, meetpunt, datum)
krw_mafa_ekr(grouped_df, watertype, taxon, aantallen)  

# of met het pipe-teken (%>%)

df_meerdere_monsters %>% 
  group_by(meetpunt, datum) %>% 
  krw_mafa_ekr(watertype, taxon, aantallen)

}