Toevoegen van de abundantieklasse

Deze functie voegt aan een dataframe een kolom abundantieklasse toe met de driedelige abundantieklassen die bij de KRW-toetsing van macrofyten wordt gebruikt. De functie kan zowel percentages als de 9-delige STOWA-klasse (afgeleid van Tansley en/of Braun-Blanquet) omzetten in de abundantieklassen.

add_abundantieklasse(df, bedekking, type = "percentage")

Arguments

df
Een dataframe met vegetatiegegevens

bedekking
Naam van de kolom met de waarden van de bedekking. Mag met en zonder "".

type
"percentage" of "klasse". Geeft aan of in de kolom bedekking percentages of klassen (1 t/m 9) staan.

Value

Het input dataframe met een toegevoegde kolom abundantieklasse

Note

Missende en/of ongeldige waarden (bijv. 0, <0 omgezet in NA. In deze gevallen geeft de functie wel een melding dat deze waarden geen correcte waarden zijn.

Examples

test <- tibble::tribble(
~naam, ~bedekkingspercentage, ~klasse,
"c",                     1,       1,
"e",                     5,       3,
"f",                    25,       4,
"g",                    49,       7,
"h",                    50,       8,
"i",                    51,       9
)

add_abundantieklasse(test, bedekkingspercentage, type = "percentage")

#> # A tibble: 6 x 4
#>   naam  bedekkingspercentage klasse abundantieklasse
#>   <chr>                <dbl>  <dbl>            <int>
#> 1 c                        1      1                1
#> 2 e                        5      3                2
#> 3 f                       25      4                2
#> 4 g                       49      7                2
#> 5 h                       50      8                2
#> 6 i                       51      9                3

add_abundantieklasse(test, "bedekkingspercentage", type = "percentage")

#> # A tibble: 6 x 4
#>   naam  bedekkingspercentage klasse abundantieklasse
#>   <chr>                <dbl>  <dbl>            <int>
#> 1 c                        1      1                1
#> 2 e                        5      3                2
#> 3 f                       25      4                2
#> 4 g                       49      7                2
#> 5 h                       50      8                2
#> 6 i                       51      9                3

add_abundantieklasse(test, klasse, type = "klasse")

#> # A tibble: 6 x 4
#>   naam  bedekkingspercentage klasse abundantieklasse
#>   <chr>                <dbl>  <dbl>            <int>
#> 1 c                        1      1                1
#> 2 e                        5      3                1
#> 3 f                       25      4                2
#> 4 g                       49      7                2
#> 5 h                       50      8                3
#> 6 i                       51      9                3